init impl of miasm features
